// services/admin-ui/src/api/schema.d.ts
export interface components {
    schemas: {
// ...
        ParticipantCreate: {
            /** Branch Id */
            branch_id: string;
            /** Prefix */
            prefix?: string | null;
            /** First Name */
            first_name: string;
            /** Last Name */
            last_name: string;
            /** Gender */
            gender?: string | null;
            /** Age */
            age?: number | null;
            /** Sub District */
            sub_district?: string | null;
            /** District */
            district?: string | null;
            /** Province */
            province?: string | null;
            /** Phone */
            phone?: string | null;
            /** Line Id */
            line_id?: string | null;
            /** Enrolled Date */
            enrolled_date?: string | null;
            /**
             * Privacy Accepted
             * @default false
             */
            privacy_accepted: boolean;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}